Deler opp skjema i flere seksjoner
